[Style] Monster, Hero 분리
