+++ Assets/LeeGyu/Scripts/Objects/AttackMethods/HiroRelated.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 480d88ca388fe75439479d6bd3af5992
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
