+++ Assets/LeeGyu/Scripts/Objects/AttackMethods/MonsterRelated.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7ebd549abf7901344a1c2bbfa13fb0f3
+guid: 1e9833e7e1d2f51408c0bf76dd20a3bf
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
