+++ Assets/LeeGyu/Scripts/Objects/AttackMethods/MonsterRelated/BaseAM.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0fc33b2d125ba64cb25bd33ac2d8f1d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
