+++ Assets/LeeGyu/Scripts/Objects/AttackMethods/MonsterRelated/ParentsAM.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fc8381f1a0e23324480ce2ea2386fdca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
