+++ Assets/LeeGyu/Scripts/Objects/AttackMethods/MonsterRelated/SkillAM.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 76291e13b2d5f5440a244cf12b97c487
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
