+++ Assets/LeeGyu/Scripts/Objects/MonsterShooter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe90f39e225386149aa18b7fa107caff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
